consomer/requirements.txt
kafka-python>=2.0.2
pandas>=1.3.3
joblib>=1.1.0
hdfs>=2.7.0
scikit-learn>=1.0.2
hdfs>=2.7.3
pyspark>=3.3.0


py4j>=0.10.9.5
requests>=2.31.0